+++ src/main/java/com/mawl/easycode/actions/ConfigAction.java
@@ -1,5 +1,6 @@
 package com.mawl.easycode.actions;
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
@@ -22,6 +23,11 @@ public class ConfigAction extends AnAction {
      */
     ConfigAction(@Nullable String text) {
         super(text);
+    }
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
     }
 
     /**