+++ src/main/java/com/mawl/easycode/actions/MainAction.java
@@ -3,6 +3,7 @@ package com.mawl.easycode.actions;
 import com.intellij.database.model.DasColumn;
 import com.intellij.database.psi.DbTable;
 import com.intellij.database.util.DasUtil;
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
@@ -45,6 +46,11 @@ public class MainAction extends AnAction {
      */
     MainAction(@Nullable String text) {
         super(text);
+    }
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
     }
 
     /**